next create level

+++ apps/webapps/admin-webapp/src/main/java/fr/k2i/adbeback/webapp/bean/GooseLevelGame.java
@@ -19,7 +19,7 @@ public class GooseLevelGame implements Serializable{
     GooseLevel level;
     List<GooseCase> gooseCases;
 
-    public Long getLevel(){
+    public Integer getLevel(){
         return level.getLevel();
     }
